1. 메뉴 api 완료 2. 권한그룹 진행중

+++ src/main/java/egovframework/com/devjitsu/repository/menu/TblMenuAuthrtGroupRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface TblMenuAuthrtGroupRepository extends JpaRepository<TblMenuAuthrtGroup, String> {
 
-    TblMenuAuthrtGroup findByAuthrtGroupSn(int authrtGroupSn);
+    TblMenuAuthrtGroup findByAuthrtGroupSn(long authrtGroupSn);
 }
