+++ src/main/java/egovframework/com/devjitsu/repository/menu/TblMenuRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface TblMenuRepository extends JpaRepository<TblMenu, String> {
 
-    TblMenu findByMenuSn(int menuSn);
+    TblMenu findByMenuSn(long menuSn);
 
     @Query(value =
             "UPDATE " +
